### 2022-02-17 #4178 @whedon
[PRE REVIEW]: Archeofrag: an R package for refitting and spatial analysis in archaeology. Presentation and tutorial
**Submitting author:** @sebastien-plutniak (<a href="http://orcid.org/0000-0002-6674-3806">Sébastien Plutniak</a>)
**Repository:** https://github.com/sebastien-plutniak/archeofrag
**Branch with paper.md** (empty if default branch):
**Version:** v0.7.1
