### 2022-02-17 #4178 @danielskatz
Assuming your submission passes this step, you should then reduce the size of your paper.  It's currently 4403 words, and JOSS suggests papers of around 1000 words.  See https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain and the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography)
