### 2022-02-27 #4178 @arfon
@kbarnhart – would you be willing to edit this submission for JOSS (providing you think this submission meets our scope)?
