### 2022-03-08 #4178 @martinfleis
Hi @sebastien-plutniak!

A few editorial comments before we'll find reviewers for your package.

1. The repository currently does not contain the `LICENSE` file with the OSI approved license under which your software is distributed. Can you please add it?
2. The paper needs to be shortened as pointed out by @danielskatz above. I suggest removing the `Use case` section and moving it to the documentation of the package.
3. The Readme (is there any other documentation?) is missing community guidelines. There should be clear guidelines for third-parties wishing to:
    - Contribute to the software
    - Report issues or problems with the software
    - Seek support

Also, if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
