+++ DESCRIPTION
@@ -1,13 +1,15 @@
 Package: archeofrag
 Type: Package
 Title: Refitting and Spatial Analysis in Archaeology
-Version: 0.7.1
-Date: 2022-02-17
+Version: 0.8
+Date: 2022-03-08
 Authors@R: person("Sebastien", "Plutniak", email = "sebastien.plutniak@posteo.net", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-6674-3806"))
 Author: Sebastien Plutniak [aut, cre] (<https://orcid.org/0000-0002-6674-3806>)
 Maintainer: Sebastien Plutniak <sebastien.plutniak@posteo.net>
 Description: Methods to analyse fragmented objects in archaeology using refitting relationships between fragments scattered in archaeological spatial units (e.g. stratigraphic layers). Graphs and graph theory are used to model archaeological observations. The package is mainly based on the 'igraph' package for graph analysis. Functions can: 1) create, manipulate, and simulate fragmentation graphs, 2) measure the cohesion and admixture of archaeological spatial units, and 3) characterise the topology of a specific set of refitting relationships. An empirical dataset is also provided as an example.
-License: GPL-2
+License: GPL-3 + file LICENSE
+Repository: CRAN
+Encoding: UTF-8
 Depends: igraph, RBGL
 Imports:
     graphics,
@@ -22,3 +24,5 @@ Suggests:
 VignetteBuilder: knitr
 URL: https://github.com/sebastien-plutniak/archeofrag
 BugReports: https://github.com/sebastien-plutniak/archeofrag/issues
+
+
