+++ vignettes/archeofrag-vignette.Rmd
@@ -1,9 +1,9 @@
 ---
-title: "Archeofrag: Tools for Refitting and Stratigraphical Analysis in Archaeology"
+title: "Archeofrag: Refitting and Spatial Analysis in Archaeology"
 author: "Sébastien Plutniak"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Archeofrag: Tools for Refitting and Stratigraphical Analysis in Archaeology}
+  %\VignetteIndexEntry{Archeofrag: Refitting and Spatial Analysis in Archaeology}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -20,7 +20,7 @@ knitr::opts_chunk$set(
 
 
 
-## Installation
+# Installation
 The development version can be installed from GitHub with:
 
 
@@ -40,343 +40,389 @@ BiocManager::install("RBGL")
 For an interactive demonstration, see also the [Shiny application](https://analytics.huma-num.fr/Sebastien.Plutniak/archeofrag/).
 
 
-- [**Introduction**](#Introduction)
-- [**Building fragmentation graphs**](#building-fragmentation-graphs): [Using empirical data](#using-empirical-data), [Generating artificial data](#generating-artificial-data)
-- [**Manipulating fragmentation graphs**](#manipulating-fragmentation-graphs)
-- [**Measuring the admixture of two layers**](#measuring-the-admixture-of-two-layers)
-- [**Characterising fragmentation within a layer**](#characterising-fragmentation-within-a-layer)
-- [**More on artificial graphs**](#more-on-artificial-graphs): [Parameters](#parameters), [Hypotheses testing](#hypotheses-testing)
+- [**Building fragmentation graphs**](#building-the-fragmentation-graph)
+- [**Edge weighting, cohesion and admixture computation**](#edge-weighting-cohesion-and-admixture-computation)
+- [**Testing layer formation hypotheses using simulated data**](#testing-layer-formation-hypotheses-using-simulated data): 
+  - [Generating artificial fragmentation graphs](#generating-artificial-fragmentation-graphs)
+  - [Testing hypotheses](#testing-hypotheses)
+- [**Assessing spatial unit boundaries using similarity relationships**](#assessing-spatial-unit-boundaries-using-similarity-relationships)
+
+- [**Characterising spatial units from their fragmentation**](#characterising-spatial-units-from-their-fragmentation)
+ 
 
 
-## Introduction 
+# Building the fragmentation graph
 
-```{r setup, message=F}
+The `Archeofrag` package comes with a small example data set called “Liang Abu”, related to the pottery fragments found on the surface and in the first two layers of the Liang Abu rock shelter. The data set contains three data frames:
+
+* a table with information about the fragments (a unique identifier, the layer, their length and width, etc.),
+* a table with the connection relationships between these fragments (each row contains the unique identifiers of two refitting fragments),
+* a table with the similarity relationships between these fragments (two fragments are termed “similar” if they seem to come from the same object but do not have connecting edges. 
+
+The `make_frag_object` function builds objects with the class “frag”. 
+Frag objects are not required by the other `Archeofrag` functions, however, using them ensures that the data are suitable for the next steps of the analysis.
+The `make_cr_graph` function takes a frag object and generates an `igraph` graph object representing the connection relationships.
+
+```{r make-empirical-graph,  message=F}
 library(archeofrag)
-```
-
-On the field, archaeologists distinguish  between two main categories of facts: layers and archaeological objects.
[…]
