+++ archeofrag-paper.bib
@@ -1,6 +1,6 @@
 @Proceedings{CzieslaEtal1990,
   Title                    = {The Big Puzzle},
-  Editor                   = {Cziesla, Erwin and Eickhoff, Sabine and Arts, Nico and Winter,  Doris},
+  Editor                   = {Cziesla, Erwin and Eickhoff, Sabine and Arts, Nico and Winter, Doris},
   Year                     = {1990},
   Eventdate                = {1987-09-06/1987-09-09},
   Number                   = {1},
@@ -14,25 +14,27 @@
 
 @Collection{SchurmansDebie2007,
   Title                    = {Fitting Rocks},
+  Subtitle                 = {Lithic Refitting Examined},
   Editor                   = {Schurmans, Utsav and De Bie, Marc},
   Year                     = {2007},
   ISBN                     = {978-1-4073-0012-2},
   Number                   = {1596},
   Publisher                = {Archaeopress},
   Series                   = {British Archaeological Reports International series},
-  Subtitle                 = {Lithic Refitting Examined},
   Address                  = {Oxford},
 }
 
  
-@Manual{Rcoreteam2020,
-    title = {R: A Language and Environment for Statistical Computing},
-    author = {{R Core Team}},
-    organization = {R Foundation for Statistical Computing},
-    address = {Vienna, Austria},
-    year = {2020},
-    url = {https://www.R-project.org/},
-  }
+@Manual{Rcoreteam2021,
+  Title                    = {R: A Language and Environment for Statistical Computing},
+  Author                   = {{R Core Team}},
+  Year                     = {2021},
+  Organization             = {R Foundation for Statistical Computing},
+  Url                      = {https://www.R-project.org/},
+  Address                  = {Vienna, Austria},
+  Langid                   = {american}
+}
+
   
   
   
@@ -75,7 +77,7 @@
   Author                   = {Philippe, Anne and  Vibet, Marie-Anne},
   Year                     = {2020},
   Volume                   = {93},
-  doi                 = {10.18637/jss.v093.c01},
+  doi                      = {10.18637/jss.v093.c01},
   Journal                  = {Journal of Statistical Software},
 }
 
@@ -120,10 +122,9 @@
   Booktitle                = {Le Taï (Remoulins – Gard). Premières sociétés agropastorales du Languedoc méditerranéen (6\textsuperscript{e}--3\textsuperscript{e} millénaire avant notre ère},
   Editor                   = {Manen, Claire},
   Publisher                = {Archives d'Écologie Préhistorique},
-  Pubstate                 = {forthcoming},
-  year = {2022},
+  pages                    = {616--687},
+  year                     = {2022},
   Address                  = {Toulouse},
-  Keywords                 = {plutniak, archeo, hceres-cha},
   Langid                   = {french}
 }
 
@@ -135,10 +136,7 @@
   ISSN                     = {0305-4403},
   Pages                    = {105501},
   Volume                   = {136},
-
[…]
