+++ archeofrag-paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Archeofrag: an R package for refitting and spatial analysis in archaeology. Presentation and tutorial'
+title: 'Archeofrag: an R package for refitting and spatial analysis in archaeology'
 tags:
   - R
   - archaeology
@@ -16,7 +16,7 @@ authors:
 affiliations:
  - name: TRACES Laboratory, Toulouse, France
    index: 1
-date: 17 February 2022 
+date: 8 March 2022 
 bibliography: archeofrag-paper.bib
 output: 
   pdf_document: 
@@ -34,9 +34,11 @@ Refitting fragments belonged to the same object at some moment in the past.
 More precisely, archaeologists deduce former connection relationships from the symmetry and the possibility of contact of significantly large surface areas from two fragments, which can be physically adjusted (the fragments “refit”).
 Here, “connection” is used as a shorthand to refer to the connection relationship that existed in the past between two areas of an object before they were broken into fragments.
 Archaeological refitting analysis has several aims: 
-1) to reconstruct objects, 
-2) to determine technical sequences (e.g., stone tool manufacture), and 
-3) to determine the reliability of spatial units and their possible admixture due to pre- and post-depositional processes. 
+
+1. to reconstruct objects, 
+2. to determine technical sequences (e.g., stone tool manufacture), and 
+3. to determine the reliability of spatial units and their possible admixture due to pre- and post-depositional processes. 
+
 This analysis has long been used for the latter aim, benefiting from multiple methodological improvements (for an overview, see @CzieslaEtal1990, @SchurmansDebie2007). 
 These methods have relied on the comparison between the number of refits between different spatial units and within these units. 
 However, it has been demonstrated that considering the number of refits without considering their topology can lead to misleading interpretations.
@@ -104,561 +106,11 @@ The archaeological interpretation of the numerical values depends on the type of
 The simulation function generates connected fragments scattered within one or two spatial units (see @Plutniak2021jas for details). It can be set with multiple parameters (number of initial objects/fragments, number of fragments, number of relationships, number of fragments and relationships, the balance between layers, etc.). It can also be constrained to generate only planar graphs since this corresponds to the fragmentation in some specific archaeological contexts (e.g., pottery with simple shapes, small sets of refits). However, the run time of this function is doubled when this constraint is used.
 
 
-# Use case: pottery from Liang Abu rock shelter
-
-## Building the fragmentation graph
-
-The `Archeofrag` package comes with a small example data set called “Liang Abu”, related to the pottery fragments found on the surface and in the first two layers of the Liang Abu rock shelter [@Plutniak2021aburefits]. The data set contains three data frames:
-
-* a table with information about the fragments (a unique identifier, the layer, their length and width, etc.),
-* a table with the connection relationships between these fragments (each row contains the unique identifiers of two refitting fragments),
-* a table with the similarity relationships between these fragments (two fragments are termed “similar” if they seem to come from the same object but do not have connecting edges. 
-
-The `make_frag_object` function builds objects with the class “frag”. 
-Frag objects are not required by the other `Archeofrag` functions, however, using them ensures that the data are suitable for the next steps of the analysis.
-The `make_cr_graph` function takes a frag object and generates an `igraph` graph object representing the connection relationships.
-
-
-```r
-library(archeofrag)
-data(LiangAbu)
-abu.frag <- make_frag_object(cr=df.cr, fragments=fragments.info)
[…]
