+++ archeofrag-paper.bib
@@ -161,3 +161,15 @@
   Journal                  = {Bulletin de la Société préhistorique française},
   Langid                   = {french}
 }
+
+
+@Manual{Plutniak2022archeofrag,
+  Title                    = {Archeofrag: an R package for Refitting and Spatial Analysis in Archaeology},
+  Author                   = {Plutniak, Sébastien},
+  Year                     = {2022},
+  Doi                      = {10.5281/zenodo.6338443},
+  Note                     = {R package version 0.8},
+  Url                      = {https://CRAN.R-project.org/package=archeofrag},
+  Abstract                 = {Methods to analyse fragmented objects in archaeology using refitting relationships between fragments scattered in archaeological spatial units (e.g. stratigraphic layers). Graphs and graph theory are used to model archaeological observations. The package is mainly based on the 'igraph' package for graph analysis. Functions can: 1) create, manipulate, and simulate fragmentation graphs, 2) measure the cohesion and admixture of archaeological spatial units, and 3) characterise the topology of a specific set of refitting relationships. An empirical dataset is also provided as an example.},
+  Langid                   = {american}
+}
