### 2022-03-08 #4178 @sebastien-plutniak
Hi @martinfleis,
thank you for handling this submission and for this first series of comments.
1. I added a license file
2. I moved the use case to the README and to the vignette of the package
3. the DESCRIPTION file includes a `BugReports` line and my email. Is this not enough?

About potential reviewers, I might suggest: platipodium, benmarwick, steko (from the JOSS reviewers list) and, additionally, Oliver Nakoinz and Alfredo Cortell.
