+++ archeofrag-paper.bib
@@ -157,7 +157,7 @@
   Year                     = {2022},
   number                   = {1},
   volume                   = {119},
-  pages                    = {},
+  pages                    = {110--113},
   Journal                  = {Bulletin de la Société préhistorique française},
   Langid                   = {french}
 }
