### 2022-03-10 #4178 @martinfleis
> the DESCRIPTION file includes a BugReports line and my email. Is this not enough?

That does not cover all what I mentioned.

```
Contribute to the software
Report issues or problems with the software
Seek support
```

I would also suggest using an open platform for support, rather than an email. It can be via Github issues as well but you don't want to answer the same questions repeatedly so the support forum should be searchable. You should also explicitly state how others could contribute to the software.
