+++ archeofrag-paper.bib
@@ -173,3 +173,13 @@
   Abstract                 = {Methods to analyse fragmented objects in archaeology using refitting relationships between fragments scattered in archaeological spatial units (e.g. stratigraphic layers). Graphs and graph theory are used to model archaeological observations. The package is mainly based on the 'igraph' package for graph analysis. Functions can: 1) create, manipulate, and simulate fragmentation graphs, 2) measure the cohesion and admixture of archaeological spatial units, and 3) characterise the topology of a specific set of refitting relationships. An empirical dataset is also provided as an example.},
   Langid                   = {american}
 }
+
+ 
+
+@Article{PlutniakEtal2022jica,
+  Title                    = {Mainland-coastal Interactions in East Borneo: Inter-site Comparison and Bayesian chronological models of two Late Pleistocene-Holocene Sequences (Liang Abu and Kimanis rock shelters)},
+  Author                   = {Plutniak, Sébastien and Araujo, Asftolfo and Agus Oktaviana, Adhi and Sugiyanto, Bambang and Chazine, Jean-Michel and Ricaut, François-Xavier},
+  Pubstate                 = {forthcoming},
+  Journal                  = {Journal of Island and Coastal Archaeology},
+  Langid                   = {american}
+}
