+++ README.Rmd
@@ -29,6 +29,10 @@ an R package for refitting and spatial analysis in archeology
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4294066.svg)](https://doi.org/10.5281/zenodo.4294066)
 
 - [**Installation**](#installation)
+
+- [**Community guidelines**](#community-guidelines)
+  - [Reporting bugs](#reporting-bugs)
+  - [Suggesting changes](#suggesting-changes)
 - [**Building fragmentation graphs**](#building-the-fragmentation-graph)
 - [**Edge weighting, cohesion and admixture computation**](#edge-weighting-cohesion-and-admixture-computation)
 - [**Testing layer formation hypotheses using simulated data**](#testing-layer-formation-hypotheses-using-simulated-data)
