+++ README.md
@@ -1,3 +1,11 @@
+---
+output: github_document
+editor_options: 
+chunk_output_type: console
+---
+
+
+
 
 
 # Archeofrag 
@@ -14,6 +22,10 @@ an R package for refitting and spatial analysis in archeology
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4294066.svg)](https://doi.org/10.5281/zenodo.4294066)
 
 - [**Installation**](#installation)
+
+- [**Community guidelines**](#community-guidelines)
+  - [Reporting bugs](#reporting-bugs)
+  - [Suggesting changes](#suggesting-changes)
 - [**Building fragmentation graphs**](#building-the-fragmentation-graph)
 - [**Edge weighting, cohesion and admixture computation**](#edge-weighting-cohesion-and-admixture-computation)
 - [**Testing layer formation hypotheses using simulated data**](#testing-layer-formation-hypotheses-using-simulated-data)
@@ -129,11 +141,11 @@ The `frag.get.layers.pair` function has additional parameters to set the minimum
 ```r
 frag.get.layers.pair(abu.g, layer.attr="layer", sel.layers=c("1", "2"),
                      size.mini=2, mixed.components.only=TRUE)
-#> IGRAPH 38f8210 UN-- 19 22 -- 
+#> IGRAPH dd872e8 UN-- 19 22 -- 
 #> + attr: frag_type (g/c), name (v/c), layer (v/c), zmin (v/n), zmax
 #> | (v/n), square (v/c), sherd.type (v/c), thickness (v/n), length (v/n),
 #> | membership (v/n), type_relation (e/c)
-#> + edges from 38f8210 (vertex names):
+#> + edges from dd872e8 (vertex names):
 #>  [1] 187--188 165--195 195--196 195--197 196--198 195--204 196--204 197--204
 #>  [9] 198--204 195--25  188--250 27 --28  27 --366 27 --367 28 --367 366--367
 #> [17] 27 --371 332--371 366--371 25 --8   28 --835 835--836
@@ -145,11 +157,11 @@ Additionally, the `frag.get.layers` function can  extract a set of specified spa
 ```r
 frag.get.layers(abu.g, layer.attr="layer", sel.layers="1")
 #> $`1`
-#> IGRAPH fbabfbf UN-- 23 18 -- 
+#> IGRAPH 0167b51 UN-- 23 18 -- 
 #> + attr: frag_type (g/c), name (v/c), layer (v/c), zmin (v/n), zmax
 #> | (v/n), square (v/c), sherd.type (v/c), thickness (v/n), length (v/n),
 #> | type_relation (e/c)
-#> + edges from fbabfbf (vertex names):
+#> + edges from 0167b51 (vertex names):
 #>  [1] 123--124  187--188  195--196  195--197  196--198  195--204  196--204 
 #>  [8] 197--204  198--204  195--25   301--302  313--314  392--408  435--441 
 #> [15] 477--478  25 --8    435--9999 441--9999
@@ -371,12 +383,12 @@ compare.res <- frag.simul.compare(abu.g12, layer.attr="layer",
                                   iter=30, summarise=FALSE)
 head(compare.res$h1.data)
 #>   edges weightsum   balance disturbance   admixture cohesion1 cohesion2
-#> 1    54  230.6982 0.2916667  0.09259259 0.024153166 0.1613814 0.8144654
-#> 2    54  226.6407 0.3611111  0.07407407 0.016336900 0.3343392 0.6493239
-#> 3    55  216.1251 0.3611111  0.12727273 0.051774593 0.3203684 0.6278570
[…]
