### 2022-03-11 #4178 @sebastien-plutniak
Ok, I added a “Community guidelines” section and the required
information in the README.

In addition, two other potential reviewers: Clemens Schmid, Néhémie
Strupler
