### 2022-05-03 #4335 @joeroe
Comment on **substantial scholarly effort**:

In the pre-review, it was noted that the repository contains a relatively small amount of code (just under 1000 lines of R) and indeed its API does boil down to a handful of core analysis functions (`frag.layers.cohesion`, `frag.layers.admixture` and `frag.simul.process`). But I'd bear in mind that this package also represents a substantial amount of methodological development (previously published by the author at https://doi.org/10.1016/j.jas.2021.105501): as outlined in the paper, refitting is an important method in archaeology but to date has only been superficially quantified. So for me, between that methodological work and the development of a polished R package to implement it, there is definitely substantial scholarly effort here.
