### 2022-05-03 sebastien-plutniak/archeofrag#1 @joeroe
CRAN check notes
Checking package installation as part of the JOSS review openjournals/joss-reviews/issues/4335.

`devtools::check()` on the latest version reports two NOTEs for the package:

- [x] `LICENSE` is not mentioned in `DESCRIPTION` (it should be `GPL-3 + file LICENSE` I believe)
- [x] `CITATION` is at the top level of the package instead of in `inst/`

```
> devtools::check()
── Building ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── archeofrag ──
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/home/joeroe/g/work/dev/archeofrag/DESCRIPTION’ ...
─  preparing ‘archeofrag’:
✔  creating vignettes (18.7s)
─  looking to see if a ‘data/datalist’ file should be added
─  building ‘archeofrag_0.8.tar.gz’

── Checking ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── archeofrag ──
Setting env vars:
• _R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
• _R_CHECK_CRAN_INCOMING_REMOTE_    : FALSE
• _R_CHECK_CRAN_INCOMING_           : FALSE
• _R_CHECK_FORCE_SUGGESTS_          : FALSE
• NOT_CRAN                          : true
── R CMD check ───────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/tmp/RtmpLs1SEN/archeofrag.Rcheck’
─  using R version 4.2.0 (2022-04-22)
─  using platform: x86_64-pc-linux-gnu (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✔  checking for file ‘archeofrag/DESCRIPTION’
─  checking extension type ... Package
─  this is package ‘archeofrag’ version ‘0.8’
─  package encoding: UTF-8
✔  checking package namespace information ...
✔  checking package dependencies (584ms)
✔  checking if this is a source package ...
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories
✔  checking for portable file names
✔  checking for sufficient/correct file permissions ...
✔  checking serialization versions
✔  checking whether package ‘archeofrag’ can be installed (2s)
✔  checking installed package size ...
✔  checking package directory ...
✔  checking for future file timestamps ...
✔  checking ‘build’ directory
N  checking top-level files ...
   File
     LICENSE
   is not mentioned in the DESCRIPTION file.
   Non-standard file/directory found at top level:
     ‘CITATION’
✔  checking for left-over files
✔  checking index information ...
N  checking package subdirectories ...
   Found the following CITATION file in a non-standard place:
     CITATION
   Most likely ‘inst/CITATION’ should be used instead.
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded ...
✔  checking whether the package can be loaded with stated dependencies ...
✔  checking whether the package can be unloaded cleanly ...
✔  checking whether the namespace can be loaded with stated dependencies ...
✔  checking whether the namespace can be unloaded cleanly ...
✔  checking loading without being on the library search path ...
✔  checking dependencies in R code (353ms)
✔  checking S3 generic/method consistency (822ms)
✔  checking replacement functions ...
✔  checking foreign function calls ...
✔  checking R code for possible problems (2.8s)
✔  checking Rd files ...
✔  checking Rd metadata ...
✔  checking Rd line widths ...
✔  checking Rd cross-references ...
✔  checking for missing documentation entries ...
✔  checking for code/documentation mismatches (853ms)
✔  checking Rd \usage sections (924ms)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking contents of ‘data’ directory ...
✔  checking data for non-ASCII characters ...
✔  checking data for ASCII and uncompressed saves ...
✔  checking installed files from ‘inst/doc’ ...
✔  checking files in ‘vignettes’ ...
✔  checking examples (2.1s)
✔  checking for unstated dependencies in vignettes ...
✔  checking package vignettes in ‘inst/doc’ ...
✔  checking re-building of vignette outputs (16.8s)
✔  checking for non-standard things in the check directory
✔  checking for detritus in the temp directory

   See
     ‘/tmp/RtmpLs1SEN/archeofrag.Rcheck/00check.log’
   for details.

── R CMD check results ─────────────────────────────────────────────────────────────────────────────────────────────────────── archeofrag 0.8 ────
Duration: 31.7s

❯ checking top-level files ... NOTE
  File
    LICENSE
  is not mentioned in the DESCRIPTION file.
  Non-standard file/directory found at top level:
    ‘CITATION’

❯ checking package subdirectories ... NOTE
  Found the following CITATION file in a non-standard place:
    CITATION
  Most likely ‘inst/CITATION’ should be used instead.

0 errors ✔ | 0 warnings ✔ | 2 notes ✖
```
