### 2022-05-03 sebastien-plutniak/archeofrag#2 @joeroe
Package should import not depend on igraph and RBGL 
Checking package installation as part of the JOSS review https://github.com/openjournals/joss-reviews/issues/4335.

`igraph` and `RBGL` are specified as `Depends` in `DESCRIPTION`, but this is generally recommended against (as explained here: https://r-pkgs.org/namespace.html#search-path).

Converting these to `Imports` and updating the code to use explicit namespaces (e.g. `igraph::graph_from_adjacency_matrix()`) would make the code more transparent and avoid unnecessarily loading all of `igraph` and `RBGL` (both big packages!) into the global namespace.
