### 2022-05-03 sebastien-plutniak/archeofrag#3 @joeroe
Is the RBGL dependency absolutely necessary?
Checking package installation as part of the JOSS review https://github.com/openjournals/joss-reviews/issues/4335...

`RBGL` is not available on CRAN, only BioConductor. This is flagged in the README and the vignette, but I still think it's not ideal. It makes installation more difficult for new users and, I don't know about you, but in ten years of using R for archaeology I've never had to install a package from BioConductor!

I'm not sure how many times functions from `RBGL` are used (#2 would help with that), but is it at all possible to e.g. substitute them for functions in a package available on CRAN, or copy the functions you need into `archaeofrag`? `RBGL` is licensed under https://opensource.org/licenses/Artistic-2.0 which I'm not familiar with though.
