### 2022-05-03 sebastien-plutniak/archeofrag#4 @joeroe
Add CRAN installation instructions to README
Checking package installation as part of the JOSS review https://github.com/openjournals/joss-reviews/issues/4335...

The package is on CRAN (https://cran.r-project.org/web/packages/archeofrag/index.html) but the README only gives instructions for installing the dev version from GitHub. I think `install.packages("archeofrag")` is much simpler for less experienced R users (read: most archaeologists!) and could be added.
