### 2022-05-03 #4335 @joeroe
Comment on **installation**/**installation instructions**:

I was able to install the package following the instructions, but I have suggested some improvements (https://github.com/sebastien-plutniak/archeofrag/issues/1, https://github.com/sebastien-plutniak/archeofrag/issues/2, https://github.com/sebastien-plutniak/archeofrag/issues/3, https://github.com/sebastien-plutniak/archeofrag/issues/4). Most notably, the package is not in R's standard repository CRAN and has a non-CRAN dependency, complicating installation for less experienced users of R.
