### 2022-05-03 sebastien-plutniak/archeofrag#5 @joeroe
No unit tests
Checking tests as part of the JOSS review https://github.com/openjournals/joss-reviews/issues/4335...

The package currently has basic automated tests from the examples in the .Rd files, run on GitHub actions.

You might consider more specific unit tests, especially if you intend to keep developing the package in a way that might break backwards compatibility. I've found [testthat](https://testthat.r-lib.org/) + [covr](https://covr.r-lib.org/) to be an effective framework for unit testing. See https://r-pkgs.org/tests.html for a guide.
