### 2022-05-03 sebastien-plutniak/archeofrag#5 @joeroe
(Apologies, I somehow completely missed that you're already running R CMD check as a GitHub Action! Edited this accordingly – should be considered an absolutely 'optional' suggestions, not something blocking the JOSS submission.)
