### 2022-05-05 sebastien-plutniak/archeofrag#1 @zoometh
For [https://github.com/openjournals/joss-reviews/issues/4335](https://github.com/openjournals/joss-reviews/issues/4335), `License`, as mentioned by @joeroe : [https://stackoverflow.com/a/56103764/2854081](https://stackoverflow.com/a/56103764/2854081)
