### 2022-05-05 sebastien-plutniak/archeofrag#2 @zoometh
Following the JOSS review https://github.com/openjournals/joss-reviews/issues/4335, confirming @joeroe note on [Depends _vs_ Imports](https://stackoverflow.com/questions/8637993/better-explanation-of-when-to-use-imports-depends#:~:text=Listing%20a%20package%20in%20either,%5B...%5D). Moving `igraph `and `RBGL `packages from `Depends `to `Imports` in the DESCRIPTION file, tagging each external function with its original package (ex: `igraph::is.igraph()`) will allow to understand which of `archaeofrag`'s functions are using the `RBGL `package
