### 2022-05-05 sebastien-plutniak/archeofrag#6 @zoometh
documentation files (.Rd)
For [https://github.com/openjournals/joss-reviews/issues/4335](https://github.com/openjournals/joss-reviews/issues/4335), typo error:

- `df.sr` and `df.cr` are mispelled in the `man/` folder: https://github.com/sebastien-plutniak/archeofrag/blob/6d8a1554617ebec3a1159342bec9e09294e71438/man/LiangAbu.Rd#L17
