### 2022-06-08 #4335 @martinfleis
Hi @sebastien-plutniak,

I see that both reviewers have opened some issues in the repository and left additional comments in this one. Can you give us an indication on when you'll get back to the package and the paper? I am just trying to ensure that the review will not get stale.

Thanks!
