### 2022-06-08 #4335 @sebastien-plutniak
Hi all,

Thank you very much for your comments and suggestions!
I had no time to revise the package in the last week, however I will
do it before the end of June. I hope it will be fine for the JOSS!
