### 2022-06-09 sebastien-plutniak/archeofrag#7 @joeroe
Improve description of the package in README and documentation
Checking "statement of need" as part of the JOSS review https://github.com/openjournals/joss-reviews/issues/4335...

I think the purpose of the package is most clearly explained in the `DESCRIPTION` file. So users can find it more easily, perhaps you could copy this to:

- [x] The top of README.md
- [ ] A package-level documentation file (i.e. `?archeofrag`). I use `usethis::use_package_doc()` to generate this automatically, but I don't think it works without roxygen.
