### 2022-07-12 sebastien-plutniak/archeofrag#7 @sebastien-plutniak
Thanks for the suggestion. I copied the description from the `DESCRIPTION` file to the `README.md`. The package also includes a vignette with further details and examples.
