### 2022-07-12 sebastien-plutniak/archeofrag#5 @sebastien-plutniak
Thanks, yes I am considering adding unit tests in future developments of this package, in addition to the examples in .Rd files and the R CMD check GitHub Action.
