### 2022-07-13 sebastien-plutniak/archeofrag#3 @sebastien-plutniak
I agree, and I have made the suggested changes:

- the `igraph` package is now in the "imports" list: the archeofrag package use numerous functions from this package (n=41), justifying to conserve the dependency. All the `igraph` functions are individually imported in the `NAMESPACE` file with `importFrom`, improving transparency.
- the `RBGL` package is now in the "suggests" list: `RBGL` is an interface to the C++ BOOST library, and only one function is used by `archeofrag` (namely, `boyerMyrvoldPlanarityTest` to test planarity — it is, to my knowledge, the only R implementation of a planarity test algorithm). Consequently,  the installation of `RBGL` (and BioConductor) is now optional, and the functionalities based on this function have been rewritten and adapted to be conditional.
