### 2022-07-13 sebastien-plutniak/archeofrag#2 @sebastien-plutniak
Thanks, all the functions imported from external packages (namely: grDevices, graphics, methods, utils, stats, igraph):
- are imported with `importFrom` in the `NAMESPACE` file, and
- are tagged in the .R files.
