### 2022-07-13 sebastien-plutniak/archeofrag#1 @sebastien-plutniak
I moved the CITATION file in inst/ as suggested.
Concerning the mention of the `LICENSE` in `DESCRIPTION`, I agree but the last time I submitted the package to the CRAN, the admin asked me to remove the "GPL-3 + file LICENSE" mention.
