+++ R/frag.simul.process.R
@@ -1,7 +1,7 @@
 .subsetsum <- function(x, target, i = 1){
   while(i != length(x)){
     s <- sum(x[1:i], na.rm = T)
-    if(s==target) break
+    if(s == target) break
     if(s > target) x[i] <- NA
     i <- i + 1
   }
@@ -11,34 +11,30 @@
 
 .select.component <- function(g, aggreg.factor){
   # high aggreg.factor favors bigger components:
-  proba <- 1/ (1 + (1:clusters(g)$no * aggreg.factor))
-  sample(order(clusters(g)$csize, decreasing = T), 1, prob = proba)
-}
-
-.connect.neighbors.if.planar <- function(g, v, v.to.add.name, neighbors){
-  g.tmp <- induced_subgraph(g, V(g)[ V(g)$object.id == V(g)[v]$object.id ])
-  g.tmp <- add_vertices(g.tmp, 1, attr = list(name = v.to.add.name))
-  g.tmp <- add_edges(g.tmp, as.character(c(rbind(v.to.add.name, neighbors)) ))
-  
-  if(boyerMyrvoldPlanarityTest(as_graphnel(g.tmp))){
-    g <- add_edges(g, c(rbind(v.to.add.name, neighbors)) )
-  }
-  g    
+  proba <- 1/ (1 + (1:igraph::components(g)$no * aggreg.factor))
+  sample(order(igraph::components(g)$csize, decreasing = T), 1, prob = proba)
 }
 
 .connect.neighbors <- function(g, v, v.to.add.name, edges, planar){
+  # if there are other neighbours than the new vertex, try to connect one of them with the new vertex.
   # identify the neighbors of the target vertex:
-  neighbors <- neighborhood(g, order = 1, nodes = v, mindist=1)
+  neighbors <- igraph::ego(g, order = 1, nodes = v, mindist = 1)
   neighbors <- unlist(neighbors)
   neighbors <- neighbors[ ! neighbors == v.to.add.name]
   neighbors <- as.character(neighbors)
-  neighbors <- sample(neighbors, sample(0:length(neighbors), 1) )
-  # if there are other neighbours than the new vertex, try connection:
-  if(length(neighbors) > 0 & (gsize(g) + length(neighbors)) < edges ){
+  # select none or one of the neighbors:
+  neighbors <- sample(neighbors, sample(0:length(neighbors), 1))
+  if(length(neighbors) > 0 & (igraph::gsize(g) + 1) < edges ){
     if(planar){
-      g <- .connect.neighbors.if.planar(g, v, v.to.add.name, neighbors)
-    }else{
-      g <- add_edges(g, c(rbind(v.to.add.name, neighbors)) ) # planarity test
+      # make a temporary version of the modified component of the graph:
+      g.tmp <- igraph::induced_subgraph(g, igraph::V(g)[ igraph::V(g)$object.id == igraph::V(g)[v]$object.id ])
+      g.tmp <- igraph::add_edges(g.tmp, as.character(c(rbind(v.to.add.name, neighbors)) ))
+      # if the modified component is still planar, then connect the new vertex and the neighbor: 
+      if(RBGL::boyerMyrvoldPlanarityTest(igraph::as_graphnel(g.tmp))){
+        g <- igraph::add_edges(g, as.character(c(rbind(v.to.add.name, neighbors)) ))
+      }
+    }else{ # if no planarity constraint:
+      g <- igraph::add_edges(g, c(rbind(v.to.add.name, neighbors)) )
     }
   }
   g
@@ -47,14 +43,15 @@
 .add.fragment <- function(g, n.components, edges, connect.neighbors, planar, aggreg.factor){
   # select component:
[…]
