+++ R/frag.edges.weighting.R
@@ -1,9 +1,9 @@
 .euclidean.distance <- function(graph, x, y, z){
-  coords <- cbind(name = V(graph)$name,
-                  x = get.vertex.attribute(graph, x),
-                  y = get.vertex.attribute(graph, y),
-                  z = get.vertex.attribute(graph, z))
-  e.list <- get.edgelist(graph)
+  coords <- cbind(name = igraph::V(graph)$name,
+                  x = igraph::vertex_attr(graph, x),
+                  y = igraph::vertex_attr(graph, y),
+                  z = igraph::vertex_attr(graph, z))
+  e.list <- igraph::as_edgelist(graph)
   colnames(e.list) <- c("name1", "name2")
   e.list <- merge(e.list, coords, by.x="name1", by.y="name", sort=FALSE)
   e.list <- merge(e.list, coords, by.x="name2", by.y="name",
@@ -27,12 +27,12 @@
 .get.morpho.spatial.params <- function(graph, x, y, z){
   # input: a graph and the names for x y z attributes
   # output: a data frame with the distance and morpho parameters for the graph edges
-  e.list <- get.edgelist(graph)
+  e.list <- igraph::as_edgelist(graph)
   e.list <- data.frame(e.list) 
   e.list$id <- 1:nrow(e.list) # a an it to sort the row later
   colnames(e.list) <- c("name1", "name2", "id")
-  morpho.df <- cbind(name = V(graph)$name, 
-                     morphometry = V(graph)$morphometry)
+  morpho.df <- cbind(name = igraph::V(graph)$name, 
+                     morphometry = igraph::V(graph)$morphometry)
   e.list <- merge(e.list, morpho.df, by.x="name1", by.y="name", sort=FALSE)
   e.list <- merge(e.list, morpho.df, by.x="name2", by.y="name",
                   suffixes=c("1", "2"), sort=FALSE)
@@ -49,7 +49,7 @@
   
   # compute spatial distances (this is done in this function to avoid repeating 3 times the command)
   e.list$distance <- 1 # default value
-  coords <- c(x, y, z)[ c(x, y, z) %in% vertex_attr_names(graph) ]
+  coords <- c(x, y, z)[ c(x, y, z) %in% igraph::vertex_attr_names(graph) ]
   if( length(coords) > 1) {   # if at least two coordinates
     e.list$distance <- .euclidean.distance(graph, x, y, z) + 1 # +1 to ensure a neutral value when all distances=0
   }
@@ -67,12 +67,12 @@
 
 .set.edge.weight <- function(g){
   #edges weight = sum of the degrees * transitivity-based index
-  if(gsize(g) == 0 )  return(0)
-  e.list <- as_edgelist(g)
+  if(igraph::gsize(g) == 0 )  return(0)
+  e.list <- igraph::as_edgelist(g)
   colnames(e.list) <- c("src", "tgt")
   v.list <- cbind(name = V(g)$name,
                   deg = igraph::degree(g),
-                  trans = transitivity(g, type = "localundirected",
+                  trans = igraph::transitivity(g, type = "localundirected",
                                        isolates = "zero", weights = NULL))
   e.list <- merge(e.list, v.list, by.x="src", by.y="name", sort=FALSE)
   e.list <- merge(e.list, v.list, by.x="tgt", by.y="name", sort=FALSE)
@@ -86,20 +86,11 @@
[…]
