+++ README.md
@@ -1,11 +1,3 @@
----
-output: github_document
-editor_options: 
-chunk_output_type: console
----
-
-
-
 
 
 # Archeofrag 
