### 2022-07-13 #4335 @sebastien-plutniak
> Hi @sebastien-plutniak,
>
> I see that both reviewers have opened some issues in the repository and left additional comments in this one. Can you give us an indication on when you'll get back to the package and the paper? I am just trying to ensure that the review will not get stale.
>
> Thanks!

Dear All,
Sorry for the long delay in revising my package following your suggestions. It is finally done!

Best,
Sébastien
