### 2022-07-17 #4335 @zoometh
@martinfleis according to me, the complete configuration of the R package ***archeofrag*** is finished, my checklist (https://github.com/openjournals/joss-reviews/issues/4335#issuecomment-1118290161) is now completed
