+++ R/frag.simul.process.R
@@ -64,6 +64,7 @@
 
 
 .main <- function(n.components, vertices, edges, balance, disturbance, aggreg.factor, planar){
+  stop.msg <- "No solution with these parameters, decrease the number of 'edges', increase the number of 'vertices', or remove one of these constraints."
   # Initialize graph:  
   g <- n.components * igraph::make_graph(c(1, 2), directed = FALSE)
   igraph::V(g)$name <- 1:igraph::gorder(g)
@@ -71,7 +72,7 @@
   # Build graph:
   if(is.infinite(vertices)){ # only edge count constraint
     if(vertices/2 < n.components){
-      stop("Increase 'vertices' or decrease 'n.components'.")}
+      stop(stop.msg)}
     while(igraph::gsize(g) < edges){
       g <- .add.fragment(g, n.components, edges, connect.neighbors=T, planar, aggreg.factor)
     }
@@ -83,14 +84,14 @@
     }
   } else if(! is.infinite(vertices) & ! is.infinite(edges) ){ # vertices & edges constraints
     if(edges < n.components + (vertices - 2 * n.components) ){
-      stop("Irrelevant parameters, decrease 'vertices' or increase 'edges'.") 
+      stop(stop.msg) 
     }   # if not enough edges to add vertices:
     if(vertices - n.components * 2 >  edges - n.components ){
-      stop("Irrelevant parameters, decrease 'vertices' or increase 'edges'.")
+      stop(stop.msg)
     }
     # if(edges > 3 * (vertices - (n.components - 1)*2) - 7 + n.components){ # uncorrect, don't work for graphs only made of dyads
     if(edges > 3 * (vertices - (n.components - 1)*2) - 6 + n.components){ # adapted version, but probably uncorrect
-      stop("Irrelevant parameters, decrease 'edges' or increase 'vertices'.")
+      stop(stop.msg)
     }  
     
     while(igraph::gorder(g) < vertices & igraph::gsize(g) < edges ){
@@ -110,9 +111,9 @@
     }
     
     if(edges - igraph::gsize(g) > e.max){
-      stop("No solution for these parameters, given the current random result. Consider removing the edge or vertex constraint.")
+      stop(stop.msg)
     }
-    while(gsize(g) < edges){
+    while(igraph::gsize(g) < edges){
       # select a component:
       selected.component <- .select.component(g, aggreg.factor)
       v.to.connect <- igraph::V(g)[ igraph::V(g)$object.id == selected.component]$name
@@ -145,11 +146,11 @@
     if(nr.v.to.disturb <= length(igraph::V(g)[ igraph::V(g)$layer == asymmetric.transport.from]) ){
       v.to.disturb <- sample(igraph::V(g)[ igraph::V(g)$layer == asymmetric.transport.from], nr.v.to.disturb)
     } else{
-      stop("The number of fragments for asymmetric transport exceeds the number of fragments in this layer. Reduce the 'disturbance' value.")
+      stop("The number of fragments for asymmetric transport exceeds the number of fragments in this layer. Decrease the 'disturbance' value.")
     }
   } 
   # reverse layer values of the selected vertices:
-  igraph::V(g)[v.to.disturb]$layer <- as.character(factor(V(g)[v.to.disturb]$layer,
+  igraph::V(g)[v.to.disturb]$layer <- as.character(factor(igraph::V(g)[v.to.disturb]$layer,
                                                   levels = c(1,2), labels = c(2,1)))
   g
 }
@@ -170,7 +171,7 @@ frag.simul.process <- function(initial.layers=2, n.components, vertices=Inf, edg
     if( ! observed.layer.attr %in% names(igraph::vertex_attr(from.observed.graph)) ){
[…]
