### 2022-07-18 e1b6ab1
fix igraph install on win
