+++ .github/workflows/r.yml
@@ -33,7 +33,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: r-lib/actions/setup-r@master
       - name: Install Dependencies (Windows)
-        run: choco install glpk pandoc
+        run: choco install glpk pandoc rtools
         if: matrix.config.os == 'windows-latest'  
       - name: Install Dependencies (Linux)
         run: sudo apt-get install  libglpk-dev libglpk40 libc6 pandoc libcurl4-openssl-dev
