### 2022-07-18 sebastien-plutniak/archeofrag#5 @sebastien-plutniak
@joeroe Finally, I have managed to add a series of unit tests and also used covr, as suggested.
