+++ codecov.yaml
@@ -2,6 +2,7 @@ comment: false
 
 ignore:
   - "R/LiangAbu.R"
+  - "R/utils.R"
 
 coverage:
   status:

[…]
