+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: archeofrag
 Type: Package
 Title: Refitting and Spatial Analysis in Archaeology
 Version: 0.8.2
-Date: 2022-07-18
+Date: 2022-07-27
 Authors@R: person("Sebastien", "Plutniak", email = "sebastien.plutniak@posteo.net", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-6674-3806"))
 Author: Sebastien Plutniak [aut, cre] (<https://orcid.org/0000-0002-6674-3806>)
 Maintainer: Sebastien Plutniak <sebastien.plutniak@posteo.net>
