+++ codecov.yaml
@@ -2,7 +2,6 @@ comment: false
 
 ignore:
   - "R/LiangAbu.R"
-  - "R/utils.R"
 
 coverage:
   status:
