### 2022-07-27 #4335 @sebastien-plutniak
@martinfleis
I have made a new tagged release ([v0.8.2](https://github.com/sebastien-plutniak/archeofrag/releases/tag/v0.8.2)),
which is archived on Zenodo (DOI: [10.5281/zenodo.6912425](https://doi.org/10.5281/zenodo.6912425))
Thank you very much for handling this submission!
