+++ archeofrag-paper.bib
@@ -36,21 +36,16 @@
 }
 
   
-  
-  
-@Manual{Plutniak2021archeofrag,
+@Manual{Plutniak2022archeofrag,
   Title                    = {Archeofrag: an R package for Refitting and Spatial Analysis in Archaeology},
   Author                   = {Plutniak, Sébastien},
-  Year                     = {2021},
-  Doi                      = {10.5281/zenodo.4720392},
-  Note                     = {R package version 0.7},
+  Year                     = {2022},
+  Doi                      = {10.5281/zenodo.6912425},
+  Note                     = {R package version 0.8.2},
   Url                      = {https://CRAN.R-project.org/package=archeofrag},
-
-  Abstract                 = {Methods to analyse fragmented objects in archaeology using refitting relationships between fragments scattered in archaeological spatial units (e.g. stratigraphic layers). Graphs and graph theory are used to model archaeological observations. The package is mainly based on the 'igraph' package for graph analysis. Functions can: 1) create, manipulate, and simulate fragmentation graphs, 2) measure the cohesion and admixture of archaeological spatial units, and 3) characterise the topology of a specific set of refitting relationships. An empirical dataset is also provided as an example.},
-  Keywords                 = {plutniak,software, hceres-ap-software},
   Langid                   = {american}
-}
-
+}  
+  
 
 
 @Article{Frerebeau2019,
@@ -119,7 +114,7 @@
 @Incollection{CaroEtal2022,
   Title                    = {Les productions céramiques du Néolithique ancien et moyen: approches céramo-stratigraphique, technologique et morpho-stylistique},
   Author                   = {Caro, Joséphine and Manen, Claire and Baux, Agathe and Plutniak, Sébastien},
-  Booktitle                = {Le Taï (Remoulins – Gard). Premières sociétés agropastorales du Languedoc méditerranéen (6\textsuperscript{e}--3\textsuperscript{e} millénaire avant notre ère},
+  Booktitle                = {Le Taï (Remoulins – Gard). Premières sociétés agropastorales du Languedoc méditerranéen (6e--3e millénaire avant notre ère},
   Editor                   = {Manen, Claire},
   Publisher                = {Archives d'Écologie Préhistorique},
   pages                    = {616--687},
@@ -161,19 +156,6 @@
   Journal                  = {Bulletin de la Société préhistorique française},
   Langid                   = {french}
 }
-
-
-@Manual{Plutniak2022archeofrag,
-  Title                    = {Archeofrag: an R package for Refitting and Spatial Analysis in Archaeology},
-  Author                   = {Plutniak, Sébastien},
-  Year                     = {2022},
-  Doi                      = {10.5281/zenodo.6338443},
-  Note                     = {R package version 0.8},
-  Url                      = {https://CRAN.R-project.org/package=archeofrag},
-  Abstract                 = {Methods to analyse fragmented objects in archaeology using refitting relationships between fragments scattered in archaeological spatial units (e.g. stratigraphic layers). Graphs and graph theory are used to model archaeological observations. The package is mainly based on the 'igraph' package for graph analysis. Functions can: 1) create, manipulate, and simulate fragmentation graphs, 2) measure the cohesion and admixture of archaeological spatial units, and 3) characterise the topology of a specific set of refitting relationships. An empirical dataset is also provided as an example.},
-  Langid                   = {american}
-}
-
  
 
 @Article{PlutniakEtal2022jica,
