### 2022-07-30 #4335 @Kevin-Mattheus-Moerman
@sebastien-plutniak

I am an AEiC for JOSS and here to help process the final steps for acceptance. I have studies this review issues, your repository, and the paper, and have some issues below that require your attention.

- [ ] Please add a plain text LICENSE file to your main repository folder and add your license text in it. Currently it looks like you just "mention" the license or link to it (e.g. https://www.r-project.org/Licenses/GPL-3) but I have not seen the actual license file. If you add a LICENSE file to your main repository folder, services like GitHub will also automatically pick up on that license. Note also that the GPL-3 license text includes a section on `How to Apply These Terms to Your New Programs`, it is recommended to follow those steps too.
- [ ] Please manually edit the ZENODO archive such that the title matches that of the paper. (@
