### 2022-07-30 #4335 @sebastien-plutniak
Hi @Kevin-Mattheus-Moerman,
Thanks for handling this submission. I made the four edits you asked.
About the correspondence between the title of the paper and the title of the Zenodo archive: the correct title is the title used for the Zenodo archive (namely: _Archeofrag: an R package for Refitting and Spatial Analysis in Archaeology_) (the words “Presentation and tutorial” must be deleted in the title of the JOSS paper).
