+++ archeofrag-paper.bib
@@ -4,6 +4,7 @@
   Year                     = {1990},
   Eventdate                = {1987-09-06/1987-09-09},
   Number                   = {1},
+  isbn                     = {3-926216-94-8},
   Publisher                = {Holos},
   Series                   = {Studies in Modern Archaeology},
   Subtitle                 = {International Symposium on Refitting Stone Artefacts},
@@ -37,7 +38,7 @@
 
   
 @Manual{Plutniak2022archeofrag,
-  Title                    = {Archeofrag: an R package for Refitting and Spatial Analysis in Archaeology},
+  Title                    = {Archeofrag: an {R} package for Refitting and Spatial Analysis in Archaeology},
   Author                   = {Plutniak, Sébastien},
   Year                     = {2022},
   Doi                      = {10.5281/zenodo.6912425},
@@ -49,7 +50,7 @@
 
 
 @Article{Frerebeau2019,
-  Title                    = {tabula: An R Package for Analysis, Seriation, and Visualization of Archaeological Count Data},
+  Title                    = {tabula: An {R} Package for Analysis, Seriation, and Visualization of Archaeological Count Data},
   Author                   = {Frerebeau, Nicolas},
   Year                     = {2019},
   Doi                      = {10.21105/joss.01821},
@@ -68,7 +69,7 @@
   }
   
   @Article{PhilippeVibet2020,
-  Title                    = {Analysis of Archaeological Phases Using the R Package ArchaeoPhases},
+  Title                    = {Analysis of Archaeological Phases Using the {R} Package ArchaeoPhases},
   Author                   = {Philippe, Anne and  Vibet, Marie-Anne},
   Year                     = {2020},
   Volume                   = {93},
@@ -89,7 +90,7 @@
 }
 
   @Manual{CareyEtal2020,
-    title = {RBGL: An interface to the BOOST graph library},
+    title = {RBGL: An interface to the {BOOST} graph library},
     author = {Carey, Vince and Long, Li and Gentleman, R.},
     year = {2020},
     note = {R package version 1.64.0},
@@ -99,7 +100,7 @@
 
 
 @Article{Plutniak2016QI,
-  Title                    = {Borneo as a Half Empty Pot: Pottery Assemblage from Liang Abu, East Kalimantan},
+  Title                    = {Borneo as a Half Empty Pot: Pottery Assemblage from {L}iang {A}bu, {E}ast {K}alimantan},
   Author                   = {Plutniak, Sébastien and Araujo, Astolfo and Puaud, Simon and Ferrié, Jean-Georges and Oktaviana, Adhi Agus and Sugiyanto, Bambang and Chazine, Jean-Michel and Ricaut, François-Xavier},
   Year                     = {2016},
   Doi                      = {10.1016/j.quaint.2015.11.080},
@@ -112,9 +113,9 @@
 
 
 @Incollection{CaroEtal2022,
-  Title                    = {Les productions céramiques du Néolithique ancien et moyen: approches céramo-stratigraphique, technologique et morpho-stylistique},
+  Title                    = {Les productions céramiques du {N}éolithique ancien et moyen: approches céramo-stratigraphique, technologique et morpho-stylistique},
   Author                   = {Caro, Joséphine and Manen, Claire and Baux, Agathe and Plutniak, Sébastien},
-  Booktitle                = {Le Taï (Remoulins – Gard). Premières sociétés agropastorales du Languedoc méditerranéen (6e--3e millénaire avant notre ère},
+  Booktitle                = {Le {T}aï ({R}emoulins – {G}ard). Premières sociétés agropastorales du Languedoc méditerranéen (6e--3e millénaire avant notre ère)},
[…]
