### 2022-07-30 #4335 @sebastien-plutniak
Thanks @Kevin-Mattheus-Moerman
I checked the pdf, and found some issues in the references. I have uploaded a corrected version of the file, could you regenerate the paper please?
