### 2022-07-31 #4335 @arfon
> @arfon https://github.com/orgs/openjournals/teams/dev can you help update the pdf? ☝️

Should be done now!
